internal/boardops: implement corner radius
Two different possibly-radiused rectangles are required:

* the board outline, consisting of 0-width wires in Dimension layer
* the copper pours in top and bottom layers, which are pulled back from
  the board outline

In Eagle, a plain old rectangle, as was previously used for the copper
pours, cannot have a corner radius. So we replace it with a polygon, ie.
we now mirror what a user would typically do when creating a copper pour
interactively in Eagle. Seems to work.